requirements.txt
streamlit
fpdf
pypdf
faiss-cpu
langchain
euriai 
langchain_community
sentence-transformers
pandas
openpyxl   # for Excel
python-docx
